Change operator for deciding when to delete the activity.
The purpose of this function appears to be to identify any activities where the contact involved and only one
other contact (at most) have been involved in the activity and to delete those activities.

Where more than 2 contacts are involved it is deemed a 'bulk activity' and the contact is removed rather the
activity deleted.

However, the operator in the query selects for deletion activities with a MINIMUM rather than MAXIMUM of
2 contacts involved. This fixes.

As an aside there is a risk that anonymising a former staff member will do scary things to activities